// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/ot/shared/dto/get-ot-personnel-type.dto.ts
export class GetOTPersonnelType_DTO {
    PersonnelTypeId: number = null;
    PersonnelType: string = null;
    IsIncentiveApplicable: boolean = false;
    CreatedOn: string = null;
    CreatedBy: number = null;
    ModifiedOn: string = null;
    ModifiedBy: number = null;
    IsActive: boolean = false;
    EmployeeRoleId: number = null;
}